fix key well angle part
I'm not sure how I ever thought it was in the right place, maybe I
moved something around...  But at any rate it has been working, so I'm
not sure whether fixing it will actually be better or worse.  It will
now act as the spring being slightly lower now, which means less
weight when pushing down and less springing up.  I'm not at all
certain that this is preferable.